More correct handling of start from FEN style positions in manacala.
Also store which fen was used to create this position originally.

If we have a position that's not the initial position but no previous
moves, then use our fen to create the next position. If we have previous
moves, and we were ALSO created with a different initial fen, then use
our original initial fen.  Otherwise back to the old mode.